separating interface
